requirements.txt
# Core Dependencies
python-dotenv==1.0.0
click==8.1.7
pyyaml==6.0.1
jsonschema==4.19.1

# Logging and Monitoring
loguru==0.7.2
colorama==0.4.6
rich==13.7.0

# Web Crawling and HTTP
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.0.7
aiohttp==3.9.1
playwright==1.40.0

# Data Processing
pandas==2.1.3
numpy>=1.26.0

# Validation and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Slack Integration
slack-sdk==3.26.1
websocket-client==1.6.4

# Security and Encryption
cryptography==41.0.7

# Process Management
psutil==5.9.6
subprocess32==3.5.4

# File Handling
watchdog==3.0.0
filelock==3.13.1

# HTML Reporting
jinja2==3.1.2
flask==3.0.0
waitress==2.1.2

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# Regex and Pattern Matching
regex==2023.10.3

# JSON Processing
ujson==5.8.0

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Progress Bars
tqdm==4.66.1
alive-progress==3.1.5

# Database (for baseline management)
sqlalchemy==2.0.23
alembic==1.12.1

# CLI Enhancement
tabulate==0.9.0
termcolor==2.3.0

# Error Tracking
sentry-sdk==1.38.0

# Configuration Management
configparser==6.0.0

# Caching
cachetools==5.3.2
redis==5.0.1

# Development Dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
ipython==8.17.2
ipdb==0.13.13